Initial updates for Re-ingest AIP
refs #6488

Reuse existing DIP from AIP MSCLs and update for newer, more general use
case.

Update determineAIPVersionKeyExitCode to default to the only path
available, as there are no exceptions to AIP processing currently.
Remove assignFauxFileUUIDs and renameDIPFauxToOrigUUIDs as we won't be
using fake IDs to separate the original SIP from the reingested AIP.

Many SQL updates:
* Fix typos, remove bizarre links
* Update microservice groups to make them more contiguous
* Verify checksums moved to after metadata processing, as metadata
generates checksums too
* Metadata reminder moved to before submission documentation and
metadata processed
* Remove old style file ID
* Update reingest chain to use existing DIP and AIP creation, instead
of custom path
* Add processingMCP to reingest path
* Remove useless magic link, clean up setup for normalization and
setting which normalization node to go to
* Delete all TasksConfigs and StandardTasksConfigs that have no foreign
keys pointing at them

+++ src/MCPClient/lib/clientScripts/restructureBagAIPToSIP.py
@@ -33,13 +33,16 @@ if __name__ == '__main__':
     
     #move everything to submission documentation
     for item in os.listdir(path):
-        if item == "metadata":
+        if item in archivematicaFunctions.OPTIONAL_FILES+["metadata"]:
             continue
         src = os.path.join(path, item)
         dst = os.path.join(path, "metadata/submissionDocumentation", item)
+        print 'Moving', src, 'to', dst
         shutil.move(src, dst)
     src = os.path.join(path, "metadata/submissionDocumentation", "data/objects")
     dst = path
-    shutil.move(src, dst)
+    if os.path.exists(src):
+        print 'Moving', src, 'to', dst
+        shutil.move(src, dst)
 
     archivematicaFunctions.create_structured_directory(path, manual_normalization=True, printing=True)
